pyproject.toml
[project]
name = "browsex"
version = "1.0.0"
description = "Multi-browser password extraction and decryption tool supporting Chrome, Brave, Edge, Opera, and Firefox"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pycryptodome>=3.20.0",
    "pyasn1>=0.6.0",
    "orjson>=3.10.0",
]

[project.optional-dependencies]
windows = [
    "pywin32>=306",
]
macos = [
    "keyring>=24.0.0",
]
linux = [
    "secretstorage>=3.3.0",
]
crossplatform = [
    "dpapick3>=0.4.0",
]
dev = [
    "pytest>=8.0.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]

[project.scripts]
browsex = "browsex.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = ["E501"]  # line too long (handled by formatter)

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
